requirements.txt
addict
diffusers
gradio
huggingface_hub
matplotlib
numpy
onnxruntime
opencv_python
Pillow
pycocotools
PyYAML
requests
setuptools
supervision
termcolor
timm
torch
torchvision
transformers
yapf
nltk
fairscale
litellm

# Core pipeline dependencies
peft>=0.7.0  # Parameter-Efficient Fine-Tuning (LoRA)
ultralytics>=8.0.0  # YOLOv8 for student models
albumentations>=1.3.0  # Advanced augmentation library
tensorboard>=2.14.0  # Training visualization
onnx>=1.14.0  # Model export
segment-anything  # SAM models
scikit-learn  # For stratified splitting
tqdm  # Progress bars
